requirements-dev.txt
flake8
isort
mypy==0.991
pytest-cov
tox
types-PyYAML
types-tabulate
